some changes

+++ app/src/main/java/com/example/weatherapp/retrofit/model/Wind.kt
@@ -7,5 +7,5 @@ data class Wind(
     @SerializedName("deg")
     val deg: Int,
     @SerializedName("speed")
-    val speed: Int
+    val speed: Double
 )
